\name{lettergen-package}
\alias{lettergen-package}
\alias{lettergen}
\docType{package}
\title{
\packageTitle{lettergen}
}
\description{
\packageDescription{lettergen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lettergen}
\packageIndices{lettergen}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lettergen}

Maintainer: \packageMaintainer{lettergen}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
